refactor: 출력 모듈 분리 -HTML템플릿생성, 렌더링
